Added Enemy/Player Health Bars and Updated Capture the Zone/Point
Created custom health bars for all enemies and first person character. Player healthbar is not yet implemented; enemy health bar visible as long as player is in range. Might change later.

Zone finally works and takes around 40 seconds for the progress bar to fill and only does so if player is in the zone and if zone progress less than 1. Have yet to implement what to do when player captures zone.